Updated after code review
